CHECK THAT SESSION COOKIES ARE ENABLED BEFORE AUTHENTICATING
I've come across a user who is blocking the session cookie. This
scenario wasn't handled gracefully. So, prior to authenticating
(processing a login request) we now check to see that we're in receipt
of a session cookie and if we're not we don't process the login and
instead inform the user that they can't use the secure fucntionality of
the website without session cookies enabled.