TRIM WHITESPACE AROUND USERID BEFORE USING IT TO CREATE LOGIN SESSION
We had an issue whereby a user was submitting a user login with
whitespace at the end of their username. It turned out that whereas
get_filtered_data trims that whitespace, and so the credentials_match
constraint was working, the query of the database to get the role to
associated with the logged in session does not and so was not finding
anything.