DON'T DISPLAY THE COOKIE ALERT IF THE SESSION COOKIE IS BEING BLOCKED
If we're not being sent the session cookie on requests then the client
is blocking them. In this case, it's pointless to display the cookie
alert and furthermore, since the cookie that we used to capture when
that has been dismissed will probably also be blocked, it would come up
at each request if we were displaying it.

As part of this I've removed the filter on which cookies are available
to the templates, so now all cookies are available within the templates.