ENABLE NEWS INSERTS INTO THE MONTHLY MAILCHIMP BULLETIN
Custom news inserts can now be inserted into the monthly Mailchimp
bulletin. They can be different for those who have registered an
interest in membership news and those who haven't. The news inserts
themselves are not tracked.

This change includes the enabling of absolute include paths in Template
Toolkit, which is something we should render unnecessary in future, and
also the deletion of the "test" copy of the bulletin's template as this
is rendered redundant by this change.