ADD LOG DISPATCH FACILITY AND FIRST USE
First use meaning, the first time it's been used to diagnose an issue.
In future, when we use this facility, we will have to add further log
statements within the application and establish a pattern for its use,
i.e. what types of things are going to be logged where and using what
log level.