RESTORE EVENTS TO TOP OF MAILCHIMP BULLETIN AND MAKE SECTION CONDITIONAL
I had moved the events section down in the bulletin recently as a
one-off for a specific bulletin. This restores it to its original
location. I plan to introduce some functionality to allow the order of
sections within the bulletin to be controlled via the command line.
Also, recently I've introduced the functionality to allow bulletins with
no events in them and so this template must change to match that and not
output an events section at all if no events are requested to be
included.