USE ADDITIONAL DATA PLUGINS IN AUTH REQUEST MODULE
This module handled the application configuration in a custom way. On
reflection this isn't necessary and it can share the
DATA::Plugin::Config module with the rest of the web application. Also,
I've incorporated the DATA::Plugin::LogDispatch module so that we can
use logging in the DATA::Auth::Request module.