SIGNIFICANT OVERHAUL OF THE APPROACH TO THE APPLICATION'S CONFIGURATION
The configuration is loaded both by the application's PSGI script and by
DATA::Plugin::Config module. I've added support for the setting of
configuration variables using environment variables, so that we can
conveniently modify things like the logging levels during ad hoc
testing. This was referred to in the POD documentation update in a prior
commit for the DATA::Plugin::Config module, so that documentation change
reflects functional change that we're now catching up on in the commit
history.

I found that the only way to use the Config::General module's ability
to use environment variables, is to load the configuration in two steps.
First the configuration is loaded using Config::General and the Perl
hash representation of the configuration is then loaded by the
Config::Context module.

The dump-conf.pl is also updated to reflect this change and to
facilitate the ability to pass the script a context.