REMOVE USE OF DEFAULT CONFIG OPTION TO CONFIG GENERAL MODULE
This was causing problems since without also specifying the option to
merge duplicates it was leading to an array of values if we tried to
override a configuration item's value in the config file with an
environment variable. Furthermore, it wasn't possibly to specify merge
duplicates as it appears that the Config::Context module depends on not
merging duplicates.